Fixed a few warnings

+++ src/aeNetServer.cpp
@@ -95,7 +95,7 @@ AetherPlayer* AetherServer_AddPlayer( AetherServerInternal* as, AetherUuid uuid 
   player->hasPendingLevelChange = false;
   
   as->priv.players.push_back( player );
-  as->pub.playerCount = as->priv.players.size();
+  as->pub.playerCount = (int32_t)as->priv.players.size();
   as->pub.allPlayers = as->priv.players.data();
   
   return player;
@@ -150,7 +150,7 @@ void AetherServer_Delete( AetherServer* _as )
 #endif
 
   ENetPeer* peers = as->priv.host->peers;
-  int32_t peerCount = as->priv.host->peerCount;
+  int32_t peerCount = (int32_t)as->priv.host->peerCount;
   for ( uint32_t i = 0; i < peerCount; i++ )
   {
     enet_peer_disconnect( &peers[ i ], 0 );
